git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/ERM.BL/Dev/Features/ActivityMigration/BLFlex;C24519

+++ Source/ApplicationServices/2Gis.Erm.BLFlex.Operations.Global/Shared/Formatter/FilsPluralizer.cs
@@ -5,7 +5,7 @@ namespace DoubleGis.Erm.BLFlex.Operations.Global.Shared.Formatter
     // Fils - арабские копейки. Множественное число (согласно википедии) - fulūs
     public sealed class FilsPluralizer : IWordPluralizer 
     {
-        public string GetPluralFor(int numberToRead)
+        public string GetPluralFor(long numberToRead)
         {
             return Math.Abs(numberToRead) == 1 ? "fils" : "fulūs";
         }
